Move all GraphQL extensions into a single place
