+++ crates/fuel-core/src/graphql_api.rs
@@ -14,10 +14,8 @@ pub mod da_compression;
 pub mod database;
 pub(crate) mod extensions;
 pub(crate) mod indexation;
-pub(crate) mod metrics_extension;
 pub mod ports;
 pub mod storage;
-pub(crate) mod validation_extension;
 pub mod worker_service;
 
 #[derive(Clone, Debug)]
