+++ crates/fuel-core/src/graphql_api/extensions.rs
@@ -1,2 +1,4 @@
 pub(crate) mod chain_state_info;
+pub(crate) mod metrics;
 pub(crate) mod required_fuel_block_height;
+pub(crate) mod validation;
