+++ crates/fuel-core/src/graphql_api/extensions/metrics.rs
